# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Bal Suraksha API
  version: "0.1.0"
  description: |
    Private child-safety support platform. All endpoints enforce organization-scoped
    authorization. Enums are fixed; unknown values are rejected server-side.
  contact:
    name: Bal Suraksha engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development

# ── Security schemes ────────────────────────────────────────────────────────
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Short-lived JWT issued on private session creation.
    staffToken:
      type: http
      scheme: bearer
      description: JWT issued on staff login (local stub for demo; OIDC in production).

  # ── Shared error envelope ─────────────────────────────────────────────────
  schemas:
...
    TipCreateRequest:
      type: object
      required: [alert_id, sighting_description]
      properties:
        alert_id:
          type: string
          format: uuid
        sighting_description:
          type: string
          maxLength: 5000
        approximate_location:
          type: string
          nullable: true
        contact_preference:
          type: string
          nullable: true
    TipCreateResponse:
      type: object
      required: [receipt_id, status]
      properties:
        receipt_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [received]

    # ── Health ────────────────────────────────────────────────────────────
...
paths:
...
  /tips:
    post:
      operationId: createTip
      summary: Submit a private sighting tip (public — opaque receipt returned)
      tags: [tips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TipCreateRequest"
      responses:
        "201":
          description: Tip received (opaque receipt only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TipCreateResponse"
  /staff/tips:
    get:
      operationId: listTips
      summary: List tips for an active alert (authorized reviewers only)
      tags: [tips]
      security:
        - staffToken: []
      parameters:
        - name: alert_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tips list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tips:
                    type: array
                    items:
                      type: object
                      required: [tip_id, sighting_description, received_at]
                      properties:
                        tip_id:
                          type: string
                          format: uuid
                        sighting_description:
                          type: string
                        approximate_location:
                          type: string
                          nullable: true
                        received_at:
                          type: string
                          format: date-time